Create ScopeofWork.md
 A Scope of Work, or, in the language of the Scrum methodology, the Epic. That is, a general statement of what the project is to achieve. That statement can be as informal as a set of bullet points in a text document, or as formal as a Requirements Document (an example can be found here).

The main thing your team has to realise is that, while the Agile methology assumes that project scope (i.e. requirements) can change, CITS3200 will, likely, be one of four units you are each doing this semester, so the team needs to be conscious of the time required. Therefore any departures from, or extensions to, the scope of work will have be carefully considered in the light of the time available to the team.